make next in BundleEntry non-atomic
A BundleEntry's next pointer is only ever set at initialization or when
cleaning up (to nullptr). In the latter case, we ensure that no readers
will use it an therefore we do not need to worry about any
inconsistencies arrising while changing it. Therefore, making it atomic
is really unnecessary.